+++ frontend/src/utils/index.ts
@@ -33,7 +33,7 @@ export function displayValue(
     case ValueType.Number:
       return number.format(value as number);
     case ValueType.Percentage:
-      return percent.format((value as number) / 100);
+      return percent.format(value as number);
     case ValueType.Date:
       return dateFormatter.format(value as Date);
     case ValueType.Copy:
